requirements.txt
torch>=2.1.0
torchvision>=0.16.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
pillow>=10.0.0
imagehash>=4.3.1
requests>=2.31.0
numpy>=1.24.0
seaborn>=0.12.0
gradio>=4.0
notebook>=7.0.0

